Add files via upload
WordHub | Chrome Extension

WordHub is a Chrome extension that allows users to easily look up the meaning of words as they browse the web. It works by providing a small search bar within the browser, where users can enter a word and get its definition. The extension is made using HTML, CSS, and JavaScript, and is designed to be lightweight and user-friendly. When a user enters a word into the search bar, the extension sends a request to an online dictionary to retrieve the definition of the word. The definition is then displayed within the browser, allowing users to easily access it without leaving the page they are currently on. WordHub is a handy tool for anyone looking to quickly look up the meanings of words while they are browsing the internet.